Restructure zvol_id
The previous implementation of zvol_id suffered from a possible
buffer overflow.  When appending the '-part' partition informtation
it was possible to overrun the provided buffer.  This issue was
resolved by dynamically allocating a buffer of the correct length
using asprintf().

Additionally, since this was such a straight forward little program
I took the opportunity to restructure the code.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Issue #569