TRIM the whole vdev on create/add/attach.
With this patch, ZFS will attempt to TRIM the whole vdev in the
following cases:

 - zpool create, zpool add, zpool attach any vdev
 - zpool import spare and L2ARC devices

We do this to make sure we start from a "clean state". This way, the
physical device knows that all the data it held before it was added to
the pool is now obsolete. This is similar to what mke2fs does.