Silence the "task trim blocked" kernel message.
Currently, the TRIM thread waits indefinitely until it is waked up by
spa_sync(). This triggers a blocked task warning in Linux if there is
no write activity on the pool for a long time.

This patch fixes the issue by making the TRIM thread wake every second
to make the kernel happy.